refactor(BridgeSize): domain/validation으로 클래스 이동
