+++ src/main/java/bridge/domain/validation/BridgeSize.java
@@ -1,4 +1,4 @@
-package bridge.domain;
+package bridge.domain.validation;
 
 public class BridgeSize {
