+++ src/test/java/bridge/domain/validation/BridgeSizeTest.java
@@ -1,7 +1,8 @@
-package bridge.domain;
+package bridge.domain.validation;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import bridge.domain.validation.BridgeSize;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
